reorganize git_metrics structure
