+++ src/git_metrics/core/runner.rs
@@ -1,7 +1,8 @@
-use crate::metrics::{
-    all_metrics, metrics_trait::Metric, result_type::MetricResultType, utils::print_results,
-};
-use crate::repo::RepoWrapper;
+use crate::git_metrics::core::repo::RepoWrapper;
+use crate::git_metrics::metrics::all_metrics;
+use crate::git_metrics::types::metrics_trait::Metric;
+use crate::git_metrics::types::result_type::MetricResultType;
+use crate::git_metrics::utils::printer::print_results;
 use gix::Repository;
 use gix::revision::Walk;
 use std::collections::HashMap;
