+++ src/git_metrics/mod.rs
@@ -1,0 +1,4 @@
+pub mod core;
+pub mod metrics;
+pub mod types;
+pub mod utils;
