+++ src/git_metrics/types/mod.rs
@@ -1,0 +1,2 @@
+pub mod metrics_trait;
+pub mod result_type;
