+++ src/git_metrics/utils/parsing.rs
@@ -1,4 +1,3 @@
-use super::{metrics_trait::Metric, result_type::MetricResultType};
 use std::collections::HashMap;
 
 pub fn parse_param_i64(params: &HashMap<String, String>, key: &str, default: i64) -> i64 {
@@ -14,14 +13,4 @@ pub fn parse_param_string<'a>(
     default: &'a String,
 ) -> &'a String {
     params.get(key).unwrap_or(default)
-}
-
-pub fn print_results(metrics: &Vec<&dyn Metric>, all_results: &HashMap<&str, MetricResultType>) {
-    for metric in metrics {
-        let name = metric.name();
-        let result = all_results.get(name).unwrap();
-        println!("--- {name} ---");
-        println!("{result:#?}");
-        println!("--- --- ---\n");
-    }
 }
